# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Boilerplate Service
...
  check_session: '/{group}/session/check',<br>
  code_verification: '/{group}/device',<br>
  device_authorization: '/{group}/device/auth',<br>
  end_session: '/{group}/session/end',<br>
  introspection: '/{group}/token/introspection',<br>
  initial_access: '/{group}/token/initial_access',<br>
  jwks: '/{group}/jwks',<br>
  pushed_authorization_request: '/{group}/request',<br>
  registration: '/{group}/reg',<br>
  revocation: '/{group}/token/revocation',<br>
  token: '/{group}/token',<br>
...
  x-logo:
    url: https://ue-platform-content.s3.amazonaws.com/0f8aeec8-8e9b-4629-b06c-6672b4a2e9ad
...
paths:
...
  '/api/{group}/account/{id}':
    get:
      tags:
        - Users
      summary: Find a User with ID
      description: Identify a single user. Will require permissions
      operationId: getAccount
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account type to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []
    delete:
      tags:
        - Users
      summary: Hard delete a user - only authGroup owner
      description: Hard delete a user - only authGroup owner
      operationId: deleteAccount
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account type to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []
    patch:
      tags:
        - Users
      summary: Patch an existing User
      description: Patch an existing User. This will require permissions.
      operationId: patchAccount
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account to update
          required: true
          schema:
            type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/jsonPatch'
        description: JSON Patch specific to your account object
        required: true

  # Invite APIs
...
  '/api/{group}/client/{id}':
    get:
      tags:
        - Clients
      summary: Sanitized Response (no secrets) - Find a client with ID
      description: Identify a single client. Will require permissions
      operationId: getOne
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account type to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []
    patch:
      tags:
        - Clients
      summary: Limited update capability from outside normal oidc api - auth group owner
      description: Limited update capability from outside normal oidc api - auth group owner
      operationId: patchOne
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account type to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/jsonPatch'
        description: JSON Patch specific to your account object
        required: true
      security:
        - bearer: []
    delete:
      tags:
        - Clients
      summary: Hard delete - authGroup owners only
      description: Hard delete - authGroup owners only
      operationId: deleteOne
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
        - name: id
          in: path
          description: id of account type to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []

  # Operation APIs for security update commands
...
  /{group}/token/initial-access:
    post:
      tags:
        - OIDC
      summary: Request an OIDC initial access token
      description: Request an OIDC initial access token. You can use this to create a client or even provide it to someone so they can register to a locked auth group (a.k.a. user pool)
      operationId: getInitialAccessToken
      parameters:
        - name: group
          in: path
          description: the auth group ID associated to your business account
          schema:
            type: string
          required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/initialAccessToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/InvalidInput'
        '417':
          $ref: '#/components/responses/ExpectationFailed'
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/initialAccessTokenRequest'
        required: false
...
servers:
  - url: http://localhost:3000
components:
  responses:
    BadRequest:
      description: Bad Request, try again
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/defaultError'
    NotFound:
      description: Requested object not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/defaultError'
    InvalidInput:
      description: Request is correct but one of your inputs is wrong
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/defaultError'
    ExpectationFailed:
      description: An expectation associated to the request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/defaultError'
    Unauthorized:
      description: Unauthorized Access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/defaultError'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 'For Bearer based tokens, simply enter the token (prefixing with "bearer" is not required). Until Swagger UI supports OIDC directly, you can use this security field with OIDC tokens as well.'
    openId:
      type: openIdConnect
      description: OIDC specific OAuth2 implementation
      openIdConnectUrl: http://example.com/group/.well-known/openid-configuration
  schemas:
    defaultError:
      required:
        - error
        - message
        - id
      properties:
        error:
          type: string
          description: Type of error
        message:
          type: string
          description: Additional information regarding the error
        id:
          type: string
          description: Unique ID to help find the error in a logging stack
...
    initialAccessTokenRequest:
      properties:
        email:
          type: string
          description: This adds an email address to the token, locking its use to a user with that email address only. useful when providing a token as part of an invite; however, this is not required in all cases. For example, you do not need this to use an IAT for OIDC client creation.
        sub:
          type: string
          description: This adds a userId to the token, locking its use to a user with that id only.
        expires:
          type: number
          description: Number of seconds after which this token expires. Default is 604800 (7 days).
    initialAccessToken:
      properties:
        iat:
          type: number
          description: issue time
        exp:
          type: number
          description: expiration time
        kind:
          type: string
          default: "InitialAccessToken"
        jti:
          type: string
          description: the token
        email:
          type: string
          format: email
          description: if provided, this is the locked user email
        sub:
          type: string
          format: uuid
          description: if provided, this is the locked user id
...
    getAccount:
      description: if the auth group is active, this is the response
      properties:
        type:
          type: string
          enum: ['Account']
        data:
          type: object
          $ref: '#/components/schemas/accountObject'
...
    getClient:
      type: object
      properties:
        type:
          type: string
          enum: ['Client']
        data:
          type: object
          $ref: '#/components/schemas/clientObject'
    writeClientObject:
      type: object
      required:
      - grant_types
      - response_types
      - redirect_uris
      properties:
        application_type:
          type: string
          default: web
          enum:
            - web
            - native
        grant_types:
          type: array
          default: []
          items:
            type: string
            enum:
             - authorization_code
             - implicit
             - refresh_token
             - client_credentials
        id_token_signed_response_alg:
          type: string
        post_logout_redirect_uris:
          type: array
          default: []
          items:
            type: string
            format: uri
        require_auth_time:
          type: boolean
          default: false
        response_types:
          type: array
          default: []
          items:
            type: string
            default: code
            enum:
              - code
              - id_token
              - token
        subject_type:
          type: string
        token_endpoint_auth_method:
          type: string
        introspection_endpoint_auth_method:
          type: string
        revocation_endpoint_auth_method:
          type: string
        client_name:
          type: string
          description: A name for the client that is unique within the authGroup
        redirect_uris:
          type: array
          default: []
          items:
            type: string
            format: uri
    clientObject:
      type: object
      required:
        - application_type
        - grant_types
        - response_types
        - client_id
        - client_name
        - client_secret
        - redirect_uris
        - auth_group
      description: this may not be all inclusive, consult the most current oidc specification
      allOf:
        - $ref: '#/components/schemas/writeClientObject'
      properties:
        request_uris:
          type: array
          default: []
          items:
            type: string
            format: uri
        client_id_issued_at:
          type: integer
        client_id:
          type: string
          description: GUID
        client_secret_expires_at:
          type: integer
          default: 0
        client_secret:
          type: string
          format: password
          description: Auto Generated Secret
        auth_group:
          type: string
          description: the tenant ID (or prettyName) for the authorization group of clients, accounts and oidc configs
        registration_client_uri:
          type: string
          description: the oidc spec client access url for api calls
          format: uri
        registration_access_token:
          type: string
          description: Required to POST or PUT updates to clients if not the authGroup owner. Rotates on each call.
...
    accountObject:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          description: guid for this account
        createdAt:
          type: string
          format: date-time
          description: when the account was created
        modifiedAt:
          type: string
          format: date-time
          description: when was the account last modified
        modifiedBy:
          type: string
          description: Id of last modifier
        authGroup:
          type: string
          description: group ID for this account. email addresses are unique to auth groups.
        verified:
          type: boolean
          description: whether or not the account was verified
        blocked:
          type: boolean
          description: root level block of the user from access to the authGroup
        active:
          type: boolean
          description: status of the account
      allOf:
        - $ref: '#/components/schemas/refWriteAccount'
    refWriteAccount:
      type: object
      properties:
        username:
          type: string
          description: optional identifier, must be unique in the authGroup. If not provided, set to email.
        email:
          type: string
          description: email address
        password:
          type: string
          description: this is a password
        txt:
          type: string
          pattern: '\+(9[976]\d|8[987530]\d|6[987]\d|5[90]\d|42\d|3[875]\d|2[98654321]\d|9[8543210]|8[6421]|6[6543210]|5[87654321]|4[987654310]|3[9643210]|2[70]|7|1)\d{1,14}$'
          description: optionally used for verifying account, requires country code. e.g. +1
        metadata:
          type: object
          description: additional context provided by the implementer
...
    jsonPatch:
      additionalProperties: false
      description: Details for JSONPatch can be found at http://jsonpatch.com/
      type: array
      items:
        type: object
        description: Reference the update model for the full paths to update
        oneOf:
          - required:
              - op
              - path
              - value
            properties:
              op:
                type: string
                enum: ['replace', 'add', 'remove', 'move', 'copy', 'test']
              path:
                type: string
                description: 'A path to the property in the data model. For example /name/firstName or /emails/-'
              value:
                type: object
                description: 'The object, array or object to set the property at the above path to'
          - required:
              - op
              - path
              - value
            properties:
              op:
                type: string
                enum: ['replace', 'add', 'remove', 'move', 'copy', 'test']
              path:
                type: string
                description: 'A path to the property in the data model. For example /name/firstName or /emails/-'
              value:
                type: string
                description: 'The string, array or object to set the property at the above path to'
          - required:
              - op
              - path
              - value
            properties:
              op:
                type: string
                enum: ['replace', 'add', 'remove', 'move', 'copy', 'test']
              path:
                type: string
                description: 'A path to the property in the data model. For example /name/firstName or /emails/-'
              value:
                type: boolean
          - required:
              - op
              - path
              - value
            properties:
              op:
                type: string
                enum: ['replace', 'add', 'remove', 'move', 'copy', 'test']
              path:
                type: string
                description: 'A path to the property in the data model. For example /name/firstName or /emails/-'
              value:
                type: integer
                description: 'The integer, array or object to set the property at the above path to'
          - required:
              - op
              - path
            properties:
              op:
                type: string
                enum: ['remove']
              path:
                type: string
                description: 'A path to the property in the data model. For example /name/firstName or /emails/-'